# Fix address retrieval in `cargo bench`