Fix address retrieval in `cargo bench`

First 200 lines from nm can contain irrelevant symbols which are all filtered out and benchmark starts testing no-op.

Change code to make sure that we always have at least 200 symbols *after* filtering out.

(I'm not sure if this was broken forever, or just depends on the distro / nm version.)

Before:
```
test context_new_and_query_location       ... bench:   3,483,195 ns/iter (+/- 176,722)
test context_new_and_query_with_functions ... bench:   3,500,515 ns/iter (+/- 171,316)
test context_new_location                 ... bench:   3,548,175 ns/iter (+/- 428,981)
test context_new_with_functions           ... bench:   3,935,070 ns/iter (+/- 1,562,079)
test context_query_location               ... bench:           0 ns/iter (+/- 0)
test context_query_with_functions         ... bench:           1 ns/iter (+/- 0)
```

After:
```
test context_new_and_query_location       ... bench:   6,120,630 ns/iter (+/- 1,106,154)
test context_new_and_query_with_functions ... bench:  37,054,340 ns/iter (+/- 4,360,871)
test context_new_location                 ... bench:   3,335,070 ns/iter (+/- 118,513)
test context_new_with_functions           ... bench:   3,335,590 ns/iter (+/- 362,814)
test context_query_location               ... bench:   1,196,900 ns/iter (+/- 252,891)
test context_query_with_functions         ... bench:   5,818,455 ns/iter (+/- 64,389)
```

Fixes #136.