@philipc 2019-09-09
Looks like this depends on distro/nm version, I was getting 6320 entries in addresses. Something further to consider is it ensure whether we are getting enough addresses that are in inlined functions, especially in `context_new_and_query_with_functions`.

BTW, I currently have some in progress performance improvements that I don't have time to finish off right now. Let me know if you plan to do any work in this area so that we don't conflict.